requirements.txt
APScheduler==3.6.3
autopep8==1.7.0
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.9.1
python-telegram-bot==13.14
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
toml==0.10.2
tornado==6.1
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.9.0
